Updated validation, fixed typos, and refactored entity configurations
This commit includes several updates and fixes. In `EmailSenderWorkerService.csproj`, a new package reference was added for data annotation-based validation. `MailSettings.cs` and `Program.cs` were updated to ensure all properties must have a value and are validated at the start of the application. A typo in `PaperCutLocalEmailSender.cs` was fixed to use the actual password from `_mailSettings.Password`. `Worker.cs` was updated to allow task cancellation. In `AirlinesDbContext.cs`, the configuration for the `AircraftsData` entity was replaced with the `AirportsData` entity. `Program.cs` under the `AirlinesApi` namespace was updated to use the actual key from `jwt.Key` and the order of service registration was changed. Lastly, the configuration for the `AircraftsData` entity was moved to its own file in `AircraftsDataConfiguration.cs` for better organization.