Added new database models and updated database connection
A new folder `Database\Base_Models\` was added to the project structure and the database connection string was updated to use `demo_airlines`. A new schema design was proposed and implemented with new classes to represent the base models of the database. The `DemoAirlinesContext` class was added to represent the DbContext of the application, including DbSet properties for each entity and methods to configure the database and entity types. The command to execute the SQL script and scaffold the DbContext were updated accordingly.